make deprecated consts an alias, and don't use internally
The consts were deprecated in favor of their counterparts in
golang.org/x/sys/unix. This patch makes them an alias / sets them
to those values, which makes it more transparent that they're the
same.

Also update internal uses of the deprecated consts, as they
should no longer be used, and updated the "stub" function
to be deprecated as well.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>