GetFromName: use /run/netns instead of /var/run/netns
The GetFromName() function looked for network namespaces in "/var/run",
whereas DeleteNamed() and NewNamed() used the "bindMountPath" const (which
points to "/run/netns"). While "/var/run" should be symlink to "/run" on
most distros, this is not a guarantee, so use the same paths so that at
least the code is consistently using this path.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>