Fix a race condition in `netns.New`
I'm working on a project that uses `netns`, and I noticed that one of our
integration tests was flaky. I traced the issue to a race in `netns.New`.
The issue is that the Go scheduler might move the current goroutine to a
different OS thread after the new namespace has been created (and joined),
but before the call to `Get()`. When this happens, the wrong namespace
handle is returned.