New functions GetFrom{Pid,Thread}WithAltProcfs
Existing functions GetFrom{Pid,Thread}() need the pid/tid argument to
be numbered as per the pid namespace of the /proc mount.

In some container configurations, this is not suitable: a container with
unshared pid namespace and the host's /proc mounted in /host/proc needs
to lookup pids numbered as per the host pid namespace in /host/proc.

GetFrom{Pid,Thread}WithAltProcfs() provide a way for such container
setups to use an alternative /proc path.