fallback to iproute bind mount path
It turns out the bindmounth path used was not the same that the one used
by iproute2, existing one is /run/netns but iproute2 uses
/var/run/netns.

It seems that in FHS 3.0, /var/run is replaced by /run; a system should
either continue to provide a /var/run directory or provide a symbolic
link from /var/run to /run for backward compatibility.

When using kubernetes pods or containers, users just mount one of the
folders but the OS will not do the symlink, so is on users the
responsability to provide the right path. To keep the same compatibility
we just implement a fallback from /run/netns to /var/run/netns so user
will have the same experience when dealing with network namespace inside
Pods or containers.

Signed-off-by: Antonio Ojea <aojea@google.com>